Fix creation of document id
After switching from wxChar to wxUniChar the creation of a document id could result in an assertion. This has been fixed by applying an appropriate type cast.

+++ src/pdfencrypt.cpp
@@ -809,7 +809,7 @@ wxPdfEncrypt::CreateDocumentId()
   int k;
   for (k = 0; k < 16; k++)
   {
-    documentId.Append(wxUniChar(id[k]));
+    documentId.Append(wxUniChar((unsigned int) id[k]));
   }
   return documentId;
 }
